std: move `sys_common::backtrace` to `sys`
