+++ src/tools/miri/tests/pass/backtrace/backtrace-global-alloc.stderr
@@ -2,8 +2,8 @@
  at $DIR/backtrace-global-alloc.rs:LL:CC
    1: <fn() as std::ops::FnOnce<()>>::call_once - shim(fn())
  at RUSTLIB/core/src/ops/function.rs:LL:CC
-   2: std::sys_common::backtrace::__rust_begin_short_backtrace
- at RUSTLIB/std/src/sys_common/backtrace.rs:LL:CC
+   2: std::sys::backtrace::__rust_begin_short_backtrace
+ at RUSTLIB/std/src/sys/backtrace.rs:LL:CC
    3: std::rt::lang_start::{closure#0}
  at RUSTLIB/std/src/rt.rs:LL:CC
    4: std::ops::function::impls::call_once
